repliy and comment question

+++ src/main/java/com/fpy/community/dto/QuestionDTO.java
@@ -3,15 +3,18 @@ package com.fpy.community.dto;
 import com.fpy.community.model.User;
 import lombok.Data;
 
+/**
+ * Created by codedrinker on 2019/5/7.
+ */
 @Data
 public class QuestionDTO {
-    private Integer id;
+    private Long id;
     private String title;
     private String description;
     private String tag;
     private Long gmtCreate;
     private Long gmtModified;
-    private Integer creator;
+    private Long creator;
     private Integer viewCount;
     private Integer commentCount;
     private Integer likeCount;
